+++ lib/include/shared_data.h
@@ -3,7 +3,7 @@
 #include <mutex>
 
 struct SharedData {
-    std::vector<int> jiffies; 
+    std::vector<unsigned long long> jiffies; 
     double percent = -1;
     std::mutex mtx;
 };
